+++ dimas/src/com/query.rs
@@ -3,15 +3,28 @@
 //! Module `query` provides an information/compute requestor `Query` which can be created using the `QueryBuilder`.
 
 // region:		--- modules
-use crate::prelude::*;
+use crate::{
+	context::ArcContext,
+	error::{DimasError, Result},
+};
 #[allow(unused_imports)]
 use std::collections::HashMap;
-use std::{fmt::Debug, marker::PhantomData, sync::Mutex};
+#[cfg(feature = "query")]
+use std::sync::RwLock;
+use std::{
+	fmt::Debug,
+	marker::PhantomData,
+	sync::{Arc, Mutex},
+	time::Duration,
+};
 use tracing::{error, instrument, Level};
 use zenoh::{
 	prelude::{sync::SyncResolve, SampleKind},
-	query::ConsolidationMode,
+	query::{ConsolidationMode, QueryTarget},
+	sample::Locality,
 };
+
+use super::message::Response;
 // endregion:	--- modules
 
 // region:		--- types
@@ -62,11 +75,14 @@ pub struct QueryBuilder<P, K, C, S>
 where
 	P: Send + Sync + Unpin + 'static,
 {
+	pub(crate) allowed_destination: Locality,
 	pub(crate) prefix: Option<String>,
+	pub(crate) timeout: Option<Duration>,
 	pub(crate) key_expr: K,
 	pub(crate) callback: C,
 	pub(crate) storage: S,
 	pub(crate) mode: ConsolidationMode,
+	pub(crate) target: QueryTarget,
 	phantom: PhantomData<P>,
 }
 
@@ -78,11 +94,14 @@ where
 	#[must_use]
 	pub const fn new(prefix: Option<String>) -> Self {
 		Self {
+			allowed_destination: Locality::Any,
 			prefix,
+			timeout: None,
 			key_expr: NoKeyExpression,
 			callback: NoResponseCallback,
 			storage: NoStorage,
 			mode: ConsolidationMode::None,
+			target: QueryTarget::BestMatching,
 			phantom: PhantomData,
 		}
 	}
@@ -92,10 +111,31 @@ impl<P, K, C, S> QueryBuilder<P, K, C, S>
 where
 	P: Send + Sync + Unpin + 'static,
 {
-	/// Set the consolidation mode
+	/// Set the [`ConsolidationMode`] of the [`Query`].
 	#[must_use]
 	pub const fn mode(mut self, mode: ConsolidationMode) -> Self {
 		self.mode = mode;
+		self
+	}
+
+	/// Set the [`QueryTarget`] of the [`Query`].
+	#[must_use]
+	pub const fn target(mut self, target: QueryTarget) -> Self {
+		self.target = target;
+		self
+	}
+
+	/// Set the allowed destination of the [`Query`].
+	#[must_use]
+	pub const fn allowed_destination(mut self, allowed_destination: Locality) -> Self {
+		self.allowed_destination = allowed_destination;
+		self
+	}
+
+	/// Set a timeout for the [`Query`].
+	#[must_use]
+	pub const fn timeout(mut self, timeout: Option<Duration>) -> Self {
+		self.timeout = timeout;
 		self
 	}
 }
@@ -108,21 +148,27 @@ where
 	#[must_use]
 	pub fn key_expr(self, key_expr: &str) -> QueryBuilder<P, KeyExpression, C, S> {
 		let Self {
+			allowed_destination,
 			prefix,
+			timeout,
 			storage,
 			callback,
 			mode,
+			target,
 			phantom,
 			..
 		} = self;
 		QueryBuilder {
+			allowed_destination,
 			prefix,
+			timeout,
 			key_expr: KeyExpression {
 				key_expr: key_expr.into(),
 			},
 			callback,
 			storage,
 			mode,
+			target,
 			phantom,
 		}
 	}
@@ -136,19 +182,25 @@ where
 			.take()
 			.map_or(topic.to_string(), |prefix| format!("{prefix}/{topic}"));
 		let Self {
+			allowed_destination,
 			prefix,
+			timeout,
 			storage,
 			callback,
 			mode,
+			target,
 			phantom,
 			..
 		} = self;
 		QueryBuilder {
+			allowed_destination,
 			prefix,
+			timeout,
 			key_expr: KeyExpression { key_expr },
 			callback,
 			storage,
 			mode,
+			target,
 			phantom,
 		}
 	}
@@ -165,20 +217,26 @@ where
 		F: FnMut(&ArcContext<P>, Response) -> Result<()> + Send + Sync + Unpin + 'static,
 	{
 		let Self {
+			allowed_destination,
 			prefix,
+			timeout,
 			key_expr,
 			storage,
 			mode,
+			target,
 			phantom,
 			..
 		} = self;
 		let callback: QueryCallback<P> = Arc::new(Mutex::new(callback));
 		QueryBuilder {
+			allowed_destination,
 			prefix,
+			timeout,
 			key_expr,
 			callback: ResponseCallback { response: callback },
 			storage,
 			mode,
+			target,
 			phantom,
 		}
 	}
@@ -196,19 +254,25 @@ where
 		storage: Arc<RwLock<std::collections::HashMap<String, Query<P>>>>,
 	) -> QueryBuilder<P, K, C, Storage<P>> {
 		let Self {
+			allowed_destination,
 			prefix,
+			timeout,
 			key_expr,
 			callback,
 			mode,
+			target,
 			phantom,
 			..
 		} = self;
 		QueryBuilder {
+			allowed_destination,
 			prefix,
+			timeout,
 			key_expr,
 			callback,
 			storage: Storage { storage },
 			mode,
+			target,
 			phantom,
 		}
 	}
@@ -223,16 +287,22 @@ where
 	///
 	pub fn build(self) -> Result<Query<P>> {
 		let Self {
+			allowed_destination,
+			timeout,
 			key_expr,
 			callback,
 			mode,
+			target,
 			..
 		} = self;
 		let key_expr = key_expr.key_expr;
 		Ok(Query {
+			allowed_destination,
+			timeout,
 			context: None,
 			key_expr,
 			mode,
+			target,
 			callback: callback.response,
 		})
 	}
@@ -266,8 +336,11 @@ pub struct Query<P>
 where
 	P: Send + Sync + Unpin + 'static,
 {
+	allowed_destination: Locality,
+	timeout: Option<Duration>,
 	pub(crate) key_expr: String,
 	mode: ConsolidationMode,
+	target: QueryTarget,
 	callback: QueryCallback<P>,
 	context: Option<ArcContext<P>>,
 }
@@ -280,6 +353,7 @@ where
 		f.debug_struct("Query")
 			.field("key_expr", &self.key_expr)
 			.field("mode", &self.mode)
+			.field("allowed_destination", &self.allowed_destination)
 			.finish_non_exhaustive()
 	}
 }
@@ -312,15 +386,25 @@ where
 	#[instrument(name="query", level = Level::ERROR, skip_all)]
 	pub fn get(&self) -> Result<()> {
 		let cb = self.callback.clone();
-		let replies = self
+		let communicator = self
 			.context
 			.clone()
-			.expect("snh")
+			.ok_or(DimasError::ShouldNotHappen)?
 			.communicator
+			.clone();
+
+		let mut query = communicator
 			.session
 			.get(&self.key_expr)
+			.target(self.target)
 			.consolidation(self.mode)
-			//.timeout(Duration::from_millis(1000))
+			.allowed_destination(self.allowed_destination);
+
+		if let Some(timeout) = self.timeout {
+			query = query.timeout(timeout);
+		};
+
+		let replies = query
 			.res_sync()
 			.map_err(|_| DimasError::ShouldNotHappen)?;
 
@@ -332,7 +416,13 @@ where
 						let guard = cb.lock();
 						match guard {
 							Ok(mut lock) => {
-								if let Err(error) = lock(&self.context.clone().expect("snh"), msg) {
+								if let Err(error) = lock(
+									&self
+										.context
+										.clone()
+										.ok_or(DimasError::ShouldNotHappen)?,
+									msg,
+								) {
 									error!("callback failed with {error}");
 								}
 							}