+++ dimas/src/com/subscriber.rs
@@ -4,35 +4,39 @@
 //! A `Subscriber` can optional subscribe on a delete message.
 
 // region:		--- modules
-use crate::{prelude::*, utils::TaskSignal};
+use crate::{
+	context::ArcContext,
+	error::{DimasError, Result},
+	utils::TaskSignal,
+};
 #[allow(unused_imports)]
 use std::collections::HashMap;
-use std::sync::{mpsc::Sender, Mutex};
+#[cfg(feature = "subscriber")]
+use std::sync::RwLock;
+use std::sync::{mpsc::Sender, Arc, Mutex};
 use tokio::task::JoinHandle;
 #[cfg(feature = "subscriber")]
 use tracing::info;
 use tracing::{error, instrument, warn, Level};
 use zenoh::{
 	prelude::{r#async::AsyncResolve, SampleKind},
+	subscriber::Reliability,
 	SessionDeclarations,
 };
+
+use super::message::Message;
 // endregion:	--- modules
 
 // region:		--- types
-/// Type definition for a subscribers `publish` callback function
+/// Type definition for a subscribers `put` callback function
 #[allow(clippy::module_name_repetitions)]
 pub type SubscriberPutCallback<P> = Arc<
-	Mutex<
-		Option<
-			Box<dyn FnMut(&ArcContext<P>, Message) -> Result<()> + Send + Sync + Unpin + 'static>,
-		>,
-	>,
+	Mutex<Box<dyn FnMut(&ArcContext<P>, Message) -> Result<()> + Send + Sync + Unpin + 'static>>,
 >;
 /// Type definition for a subscribers `delete` callback function
 #[allow(clippy::module_name_repetitions)]
-pub type SubscriberDeleteCallback<P> = Arc<
-	Mutex<Option<Box<dyn FnMut(&ArcContext<P>) -> Result<()> + Send + Sync + Unpin + 'static>>>,
->;
+pub type SubscriberDeleteCallback<P> =
+	Arc<Mutex<Box<dyn FnMut(&ArcContext<P>) -> Result<()> + Send + Sync + Unpin + 'static>>>;
 // endregion:	--- types
 
 // region:		--- states
@@ -80,6 +84,7 @@ where
 	pub(crate) key_expr: K,
 	pub(crate) put_callback: C,
 	pub(crate) storage: S,
+	pub(crate) reliability: Reliability,
 	pub(crate) delete_callback: Option<SubscriberDeleteCallback<P>>,
 }
 
@@ -95,6 +100,7 @@ where
 			key_expr: NoKeyExpression,
 			put_callback: NoPutCallback,
 			storage: NoStorage,
+			reliability: Reliability::BestEffort,
 			delete_callback: None,
 		}
 	}
@@ -104,6 +110,13 @@ impl<P, K, C, S> SubscriberBuilder<P, K, C, S>
 where
 	P: Send + Sync + Unpin + 'static,
 {
+	/// Set reliability
+	#[must_use]
+	pub const fn set_reliability(mut self, reliability: Reliability) -> Self {
+		self.reliability = reliability;
+		self
+	}
+
 	/// Set liveliness subscribers callback for `delete` messages
 	#[must_use]
 	pub fn delete_callback<F>(self, callback: F) -> Self
@@ -115,15 +128,17 @@ where
 			key_expr,
 			put_callback,
 			storage,
+			reliability,
 			..
 		} = self;
 		let delete_callback: Option<SubscriberDeleteCallback<P>> =
-			Some(Arc::new(Mutex::new(Some(Box::new(callback)))));
+			Some(Arc::new(Mutex::new(Box::new(callback))));
 		Self {
 			prefix,
 			key_expr,
 			put_callback,
 			storage,
+			reliability,
 			delete_callback,
 		}
 	}
@@ -133,7 +148,7 @@ impl<P, C, S> SubscriberBuilder<P, NoKeyExpression, C, S>
 where
 	P: Send + Sync + Unpin + 'static,
 {
-	/// Set the full expression for the subscriber
+	/// Set the full key expression for the [`Subscriber`].
 	#[must_use]
 	pub fn key_expr(self, key_expr: &str) -> SubscriberBuilder<P, KeyExpression, C, S> {
 		let Self {
@@ -141,6 +156,7 @@ where
 			storage,
 			put_callback,
 			delete_callback,
+			reliability,
 			..
 		} = self;
 		SubscriberBuilder {
@@ -150,12 +166,13 @@ where
 			},
 			put_callback,
 			storage,
+			reliability,
 			delete_callback,
 		}
 	}
 
-	/// Set only the message qualifing part of the subscriber.
-	/// Will be prefixed with agents prefix.
+	/// Set only the message qualifing part of the [`Subscriber`].
+	/// Will be prefixed with [`Agent`]s prefix.
 	#[must_use]
 	pub fn topic(mut self, topic: &str) -> SubscriberBuilder<P, KeyExpression, C, S> {
 		let key_expr = self
@@ -166,6 +183,7 @@ where
 			prefix,
 			storage,
 			put_callback,
+			reliability,
 			delete_callback,
 			..
 		} = self;
@@ -174,6 +192,7 @@ where
 			key_expr: KeyExpression { key_expr },
 			put_callback,
 			storage,
+			reliability,
 			delete_callback,
 		}
 	}
@@ -193,15 +212,17 @@ where
 			prefix,
 			key_expr,
 			storage,
+			reliability,
 			delete_callback,
 			..
 		} = self;
-		let callback: SubscriberPutCallback<P> = Arc::new(Mutex::new(Some(Box::new(callback))));
+		let callback: SubscriberPutCallback<P> = Arc::new(Mutex::new(Box::new(callback)));
 		SubscriberBuilder {
 			prefix,
 			key_expr,
 			put_callback: PutCallback { callback },
 			storage,
+			reliability,
 			delete_callback,
 		}
 	}
@@ -222,6 +243,7 @@ where
 			prefix,
 			key_expr,
 			put_callback,
+			reliability,
 			delete_callback,
 			..
 		} = self;
@@ -230,6 +252,7 @@ where
 			key_expr,
 			put_callback,
 			storage: Storage { storage },
+			reliability,
 			delete_callback,
 		}
 	}
@@ -239,20 +262,22 @@ impl<P, S> SubscriberBuilder<P, KeyExpression, PutCallback<P>, S>
 where
 	P: Send + Sync + Unpin + 'static,
 {
-	/// Build the subscriber
-	/// # Errors
+	/// Build the [`Subscriber`].
 	///
+	/// # Errors
+	/// Currently none
 	pub fn build(self) -> Result<Subscriber<P>> {
 		let Self {
 			key_expr,
 			put_callback,
+			reliability,
 			delete_callback,
 			..
 		} = self;
-		dbg!(&key_expr.key_expr);
 		Ok(Subscriber {
 			key_expr: key_expr.key_expr,
-			put_callback: Some(put_callback.callback),
+			put_callback: put_callback.callback,
+			reliability,
 			delete_callback,
 			handle: None,
 		})
@@ -264,9 +289,10 @@ impl<P> SubscriberBuilder<P, KeyExpression, PutCallback<P>, Storage<P>>
 where
 	P: Send + Sync + Unpin + 'static,
 {
-	/// Build and add the subscriber to the agent
-	/// # Errors
+	/// Build and add the [`Subscriber`] to the [`Agent`].
 	///
+	/// # Errors
+	/// Currently none
 	#[cfg_attr(any(nightly, docrs), doc, doc(cfg(feature = "subscriber")))]
 	pub fn add(self) -> Result<Option<Subscriber<P>>> {
 		let c = self.storage.storage.clone();
@@ -288,7 +314,8 @@ where
 	P: Send + Sync + Unpin + 'static,
 {
 	key_expr: String,
-	put_callback: Option<SubscriberPutCallback<P>>,
+	put_callback: SubscriberPutCallback<P>,
+	reliability: Reliability,
 	delete_callback: Option<SubscriberDeleteCallback<P>>,
 	handle: Option<JoinHandle<()>>,
 }
@@ -318,45 +345,43 @@ where
 		drop(tx);
 
 		{
-			if let Some(pcb) = self.put_callback.clone() {
-				if let Err(err) = pcb.lock() {
-					warn!("found poisoned put Mutex");
-					self.put_callback
-						.replace(Arc::new(Mutex::new(err.into_inner().take())));
+			if self.put_callback.lock().is_err() {
+				warn!("found poisoned put Mutex");
+				self.put_callback.clear_poison();
+			}
+			{
+				if let Some(dcb) = self.delete_callback.clone() {
+					if dcb.lock().is_err() {
+						warn!("found poisoned delete Mutex");
+						dcb.clear_poison();
+					}
 				}
 			}
-		}
-		{
-			if let Some(dcb) = self.delete_callback.clone() {
-				if let Err(err) = dcb.lock() {
-					warn!("found poisoned delete Mutex");
-					self.delete_callback
-						.replace(Arc::new(Mutex::new(err.into_inner().take())));
-				}
-			}
-		}
 
-		let key_expr = self.key_expr.clone();
-		let p_cb = self.put_callback.clone();
-		let d_cb = self.delete_callback.clone();
+			let key_expr = self.key_expr.clone();
+			let p_cb = self.put_callback.clone();
+			let d_cb = self.delete_callback.clone();
+			let reliability = self.reliability;
 
-		self.handle
-			.replace(tokio::task::spawn(async move {
-				#[cfg(feature = "subscriber")]
-				let key = key_expr.clone();
-				std::panic::set_hook(Box::new(move |reason| {
-					error!("subscriber panic: {}", reason);
+			self.handle
+				.replace(tokio::task::spawn(async move {
 					#[cfg(feature = "subscriber")]
-					if let Err(reason) = tx.send(TaskSignal::RestartSubscriber(key.clone())) {
-						error!("could not restart subscriber: {}", reason);
-					} else {
-						info!("restarting subscriber!");
+					let key = key_expr.clone();
+					std::panic::set_hook(Box::new(move |reason| {
+						error!("subscriber panic: {}", reason);
+						#[cfg(feature = "subscriber")]
+						if let Err(reason) = tx.send(TaskSignal::RestartSubscriber(key.clone())) {
+							error!("could not restart subscriber: {}", reason);
+						} else {
+							info!("restarting subscriber!");
+						};
+					}));
+					if let Err(error) = run_subscriber(key_expr, p_cb, d_cb, reliability, ctx).await
+					{
+						error!("spawning subscriber failed with {error}");
 					};
 				}));
-				if let Err(error) = run_subscriber(key_expr, p_cb, d_cb, ctx).await {
-					error!("spawning subscriber failed with {error}");
-				};
-			}));
+		}
 	}
 
 	/// Stop a running Subscriber
@@ -371,8 +396,9 @@ where
 #[instrument(name="subscriber", level = Level::ERROR, skip_all)]
 async fn run_subscriber<P>(
 	key_expr: String,
-	p_cb: Option<SubscriberPutCallback<P>>,
+	p_cb: SubscriberPutCallback<P>,
 	d_cb: Option<SubscriberDeleteCallback<P>>,
+	reliability: Reliability,
 	ctx: ArcContext<P>,
 ) -> Result<()>
 where
@@ -382,6 +408,7 @@ where
 		.communicator
 		.session
 		.declare_subscriber(&key_expr)
+		.reliability(reliability)
 		.res_async()
 		.await
 		.map_err(|_| DimasError::ShouldNotHappen)?;
@@ -395,31 +422,27 @@ where
 		match sample.kind {
 			SampleKind::Put => {
 				let msg = Message(sample);
-				if let Some(cb) = p_cb.clone() {
-					let result = cb.lock();
-					match result {
-						Ok(mut cb) => {
-							if let Err(error) = cb.as_deref_mut().expect("snh")(&ctx, msg) {
-								error!("put callback failed with {error}");
-							}
+				match p_cb.lock() {
+					Ok(mut lock) => {
+						if let Err(error) = lock(&ctx, msg) {
+							error!("subscriber put callback failed with {error}");
 						}
-						Err(err) => {
-							error!("put callback lock failed with {err}");
-						}
+					}
+					Err(err) => {
+						error!("subscriber put callback lock failed with {err}");
 					}
 				}
 			}
 			SampleKind::Delete => {
 				if let Some(cb) = d_cb.clone() {
-					let result = cb.lock();
-					match result {
-						Ok(mut cb) => {
-							if let Err(error) = cb.as_deref_mut().expect("snh")(&ctx) {
-								error!("delete callback failed with {error}");
+					match cb.lock() {
+						Ok(mut lock) => {
+							if let Err(error) = lock(&ctx) {
+								error!("subscriber delete callback failed with {error}");
 							}
 						}
 						Err(err) => {
-							error!("delete callback lock failed with {err}");
+							error!("subscriber delete callback lock failed with {err}");
 						}
 					}
 				}