+++ dimas/src/prelude.rs
@@ -3,18 +3,24 @@
 //! Public interface of dimas. Typically it is sufficient to include the prelude with
 //! ```use dimas::prelude::*;```
 
+pub extern crate bitcode;
+
 // region:    --- modules
 // re-exports
-//pub use anyhow::Result;
-// anyhows Result
 // used std synchronisation primitives
 pub use std::sync::Arc;
 pub use std::sync::RwLock;
 // bitcode encoding/decoding
-pub(crate) use bitcode::{decode, encode};
 pub use bitcode::{Decode, Encode};
-pub extern crate bitcode;
+// zenoh stuff
+pub use zenoh::publication::CongestionControl;
+pub use zenoh::publication::Priority;
+pub use zenoh::query::ConsolidationMode;
+pub use zenoh::query::QueryTarget;
+pub use zenoh::sample::Locality;
+pub use zenoh::subscriber::Reliability;
 
+// dimas stuff
 pub use crate::agent::Agent;
 pub use crate::com::liveliness_subscriber::{LivelinessSubscriber, LivelinessSubscriberBuilder};
 pub use crate::com::message::{Message, Request, Response};
@@ -26,6 +32,7 @@ pub use crate::config::Config;
 pub use crate::context::ArcContext;
 pub use crate::error::{DimasError, Result};
 pub use crate::timer::{Timer, TimerBuilder};
+pub use crate::utils::init_tracing;
 // endregeion:  --- modules
 
 // region:    --- types