+++ examples/pubsub/src/subscriber.rs
@@ -26,8 +26,8 @@ fn hello_deletion(ctx: &ArcContext<AgentProps>) -> Result<()> {
 
 #[tokio::main]
 async fn main() -> Result<()> {
-	// a tracing subscriber writing logs
-	tracing_subscriber::fmt::init();
+	// initialize tracing/logging
+	init_tracing();
 
 	// create & initialize agents properties
 	let properties = AgentProps { test: 0 };