+++ examples/queries/src/queryable.rs
@@ -23,8 +23,8 @@ fn queryable(ctx: &ArcContext<AgentProps>, request: Request) -> Result<()> {
 
 #[tokio::main]
 async fn main() -> Result<()> {
-	// a tracing subscriber writing logs
-	tracing_subscriber::fmt::init();
+	// initialize tracing/logging
+	init_tracing();
 
 	// create & initialize agents properties
 	let properties = AgentProps { counter: 0 };